[+FEATURE] Add support for varnish running on different port
By default purge requests are made to the URL of the site. But som (many)
installations have Apache running on port 80, and Varnish running on a different
port (typocally 8080) and then in a firewall forwarding port 80 to Varnish port
which then proxies to Apache on port 80.

This means that purge request from the webserver will hit the webserver itself,
and not Varnish.

This change introduces a new option to allow specifying which port Varnish is running
on, and that port will be used for PURGE requests.

Change-Id: I522d1b8c1ba0fbb65c81581ea418393248507e9a