[FEATURE][TASK] Add wildcard appending option & improve some purge features. Code cleanup.
Adds a extension configuration to append wildcard (.*) to all url's when clearing single pages, but not when it's the frontpage.

Fixed an issue when clearing single page cache for rootpages, since they don't have an entry in the realurl pathcache.

Made cache clearing for all pages (.*) work with all domains configured in realurl configuration.

Change-Id: I1160eeb0732106b14e6b20d2fb834cd089ae99cf