[!!!][FEATURE] Add cache clearing of multiple varnish hosts
Added possibility to configure a list of Varnish hosts to purge. This enables
setups with more than one Varnish host, and setups where the Varnish server is
running on another domain/port seen from the backend webservers point of view.

This is a breaking change because it removes the overrideVarnishPort configuration.
Instead use the new varnishHosts configuration which can take servername:port for instance
varnishserver.example.com:8080

Change-Id: I5cec157701bcbfaf853be24f223182982a6480ac
Resolves: #46874