[!!!][FEATURE] Rewrite to 6.0 compatibility
The extension now has minimum 6.0 requirement.

The ESI features are no longer part of the extensions, so its no longer possible
to automatically convert USER_INT objects to edge-side-includes.

The options in the extensionmanager are changed, and should be re-configures when updating.

Complete restructuring using a simple eventbased approach.

Requires DB updates

Change-Id: Ib758882d3fae3755c9ebcaf2d740adef6093b65c
Refs: #46875