[!!!][FEATURE] Refactor how TYPO3 pages are purged in varnish
This is a conceptual difference. Inspired by Snowflakes varnish extension
TYPO3 now sends a header with the TYPO3 page uid when rendering pages. This
is used when purging/banning in varnish.

Instead of banning a specific URL, we ban objects stored in varnish with the
TYPO3 page id. This should make cache-clearing more robust when pages are for
some reason not in the RealURL patchcache (which was until now used to lookup
URLs).

This commit changes how Varnish should be configured. It still uses the custom
PURGE requests for purging single URL's (for backwards compatiblilty reasons), but
 a new custom BAN requests is introduced which bans for a given TYPO3 page uid, or
bans all.

Use this VCL snippet in you vcl_fetch

 if(req.http.Varnish-Ban-All) {
   ban("req.url ~ / && req.http.host == " + req.http.host);
   error 200 "Banned all";
 }
 if(req.http.Varnish-Ban-TYPO3-Pid) {
   ban("obj.http.TYPO3-Pid == " + req.http.Varnish-Ban-TYPO3-Pid + " && req.http.host == " + req.http.host);
   error 200 "Banned TYPO3 pid " + req.http.Varnish-Ban-TYPO3-Pid;
 }

This removed the dependency on realurl.

This makes the extension

Change-Id: Ida2690bf9b842766491f212343b6f770a6166b2c